+++ Scripts/Components/HealthPointsComponent.cs
@@ -31,6 +31,14 @@ namespace ShiftingSepulcher
 
         public void TakeDamage(int damage)
         {
+            if (IsInvulnerable)
+                return;
+
+            TakeDamageDisregardingInvulnerability(damage);
+        }
+
+        public void TakeDamageDisregardingInvulnerability(int damage)
+        {
             Health -= damage;
             _cooldownTimer = InvulnerabilityTime;
 
@@ -40,8 +48,15 @@ namespace ShiftingSepulcher
 
         public void OnTookDamageFromHitBox(HitBox hitBox)
         {
+            if (IsInvulnerable)
+                return;
+
             TakeDamage(hitBox.Damage);
+
             EmitSignal(nameof(TookDamageFromHitBox), hitBox);
+
+            hitBox.CallDeferred("emit_signal", nameof(HitBox.DealtDamageTo), this);
+            hitBox.CallDeferred("emit_signal", nameof(HitBox.DealtDamageNoParams));
         }
 
         public Vector2 GetRecoilVelocity(Vector2 attackerGlobalPosition, float friction)